+++ server/src/main/java/io/craigmiller160/orgbuilder/server/data/jdbc/AddressDao.java
@@ -203,13 +203,13 @@ public class AddressDao extends AbstractJdbcDao<AddressDTO,Long> {
     }
 
     @Override
-    public int getCount() throws OrgApiDataException {
+    public long getCount() throws OrgApiDataException {
         OrgApiLogger.getDataLogger().trace("Address Count Query:\n" + COUNT_QUERY);
-        int count = -1;
+        long count = -1;
         try(Statement stmt = getConnection().createStatement()){
             try(ResultSet resultSet = stmt.executeQuery(COUNT_QUERY)){
                 if(resultSet.next()){
-                    count = resultSet.getInt("address_count");
+                    count = resultSet.getLong("address_count");
                 }
             }
         }
