+++ server/src/main/java/io/craigmiller160/orgbuilder/server/data/jdbc/OrgDao.java
@@ -152,13 +152,13 @@ public class OrgDao extends AbstractJdbcDao<OrgDTO,Long> {
     }
 
     @Override
-    public int getCount() throws OrgApiDataException {
+    public long getCount() throws OrgApiDataException {
         OrgApiLogger.getDataLogger().trace("Org Count Query:\n" + COUNT_QUERY);
-        int count = -1;
+        long count = -1;
         try(Statement stmt = getConnection().createStatement()){
             try(ResultSet resultSet = stmt.executeQuery(COUNT_QUERY)){
                 if(resultSet.next()){
-                    count = resultSet.getInt("org_count");
+                    count = resultSet.getLong("org_count");
                 }
             }
         }
